pyproject.toml
[tool.poetry]
name = "jira-agent"
version = "0.1.0"
description = "Autonomous Jira ticket agent using Claude Agent SDK"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "jira_agent"}]

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.40.0"
requests = "^2.32.0"
pyyaml = "^6.0.2"
docopt = "^0.6.2"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
keyring = "^25.5.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
httpx = "^0.28.0"
gitpython = "^3.1.43"
databricks-sdk = "^0.38.0"
databricks-sql-connector = "^3.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.0"
ruff = "^0.8.0"
responses = "^0.25.0"

[tool.poetry.scripts]
jira-agent = "jira_agent.main:main"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
